agregando manejo de errores para paciente y Turnos

+++ src/main/java/org/umsaback/services/PacienteService.java
@@ -1,10 +1,10 @@
 package org.umsaback.services;
 import java.util.List;
 
+import org.umsaback.exceptions.PatientNotFoundException;
 import org.umsaback.models.dtos.PacienteDTO;
 import org.umsaback.models.entities.Paciente;
 import org.umsaback.repositories.PacienteRepository;
-
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 
@@ -21,9 +21,17 @@ public class PacienteService {
 			return pacienteRepository.listAll();
 		}
 	 
-	 public Paciente findByCuit(String cuit){
+	 //NullPointerException para valores null ?
+	 
+	 public Paciente findByCuit(String cuit) throws PatientNotFoundException {
 		 
-			return pacienteRepository.findByCuit(cuit);
+		 	Paciente paciente = pacienteRepository.findByCuit(cuit);
+		 
+		    if (paciente == null) {
+		    	
+		        throw new PatientNotFoundException("Patient with CUIT " + cuit + " not found");
+		    }
+			return paciente;
 		}	 
 	 
 	 
@@ -34,16 +42,21 @@ public class PacienteService {
 		 			pacienteDTO.getDomicilio(),pacienteDTO.getFechaNacimiento(),pacienteDTO.getCelular(),
 		 			pacienteDTO.getObraSocial(),pacienteDTO.getCuit()
 		 			);
-		
+			 	
 		 	pacienteRepository.persist(nuevoPaciente);
 		
 	        return pacienteRepository.findById(nuevoPaciente.getId());
 	    }
 	 
 	
-	 public Paciente deletePacienteByCuit(String cuit) {
+	 
+	 public Paciente deletePacienteByCuit(String cuit) throws PatientNotFoundException{
 		 
 		 Paciente deletedPatient = pacienteRepository.findByCuit(cuit);
+		 
+		 if(deletedPatient == null) {
+			 throw new PatientNotFoundException("Patient with CUIT " + cuit + " not found");
+		 }
 		 
 		 pacienteRepository.delete(deletedPatient);
 		 
@@ -51,12 +64,17 @@ public class PacienteService {
 	 }
 	 
 	 
-	 public Paciente updatePatient(String cuit, PacienteDTO pacienteDTO ) {
+	 public Paciente updatePatient(String cuit, PacienteDTO pacienteDTO ) throws PatientNotFoundException{
 		 
 		 Paciente updatedPatient = pacienteRepository.findByCuit(cuit);
 		 
+		 if(updatedPatient == null) {
+			 throw new PatientNotFoundException("Patient with CUIT " + cuit + " not found");
+		 }
+		 
 		 // Para el caso de un solo campo ?
 	     //pacienteRepository.getEntityManager().merge(pacienteDTO);
+		 // Debe haber otro metodo para hacer un merge
 		 
 		 updatedPatient.setNombre(pacienteDTO.getNombre());
 		 updatedPatient.setApellido(pacienteDTO.getApellido());
